Swagger-UI

+++ src/main/java/com/matchvagas/backend/dto/EstadoRequestDTO.java
@@ -1,7 +1,3 @@
 package com.matchvagas.backend.dto;
 
-public record EstadoRequestDTO(
-    String nome,
-    String uf,
-    Integer paisId
-) {}
+public record EstadoRequestDTO(String nome, String uf, Long paisId) {}
