+++ src/main/java/com/matchvagas/backend/dto/PaisResponseDTO.java
@@ -3,8 +3,4 @@ package com.matchvagas.backend.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record PaisResponseDTO(
-    Integer id,
-    String nome,
-    String codigoIso
-) {}
+public record PaisResponseDTO(Long id, String nome, String codigoIso) {}
